requirements.txt
twilio
python-dotenv
customtkinter
pillow